pyproject.toml

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "unetloopdetection"
version = "0.1.0"
description = "U-Net utilities for loop detection"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "Hadi"}]
dependencies = [
  "numpy>=1.23",
  "torch>=2.1",
  "torchvision>=0.16",
  "tqdm>=4.66",
  "pyyaml>=6.0",
  "matplotlib>=3.7",
  "pandas>=2.0",
  "scikit-learn>=1.3",
  "imageio>=2.31",
  "pillow>=10.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
